+++ R/mr.R
@@ -118,7 +118,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_simple_median",
 			name="Simple median",
-			PubmedID="",
+			PubmedID="27061298",
 			Description="",
 			use_by_default=FALSE,
 			heterogeneity_test=FALSE
@@ -126,7 +126,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_weighted_median",
 			name="Weighted median",
-			PubmedID="",
+			PubmedID="27061298",
 			Description="",
 			use_by_default=TRUE,
 			heterogeneity_test=FALSE
@@ -134,7 +134,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_penalised_weighted_median",
 			name="Penalised weighted median",
-			PubmedID="",
+			PubmedID="27061298",
 			Description="",
 			use_by_default=FALSE,
 			heterogeneity_test=FALSE
@@ -142,7 +142,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_ivw",
 			name="Inverse variance weighted",
-			PubmedID="",
+			PubmedID="24114802",
 			Description="",
 			use_by_default=TRUE,
 			heterogeneity_test=TRUE
@@ -150,7 +150,7 @@ mr_method_list <- function()
 		list(
 			obj = "mr_ivw_radial",
 			name = "IVW radial",
-			PubmedID = "",
+			PubmedID = "29961852",
 			Description = "",
 			use_by_default = FALSE,
 			heterogeneity_test = TRUE
@@ -174,7 +174,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_simple_mode",
 			name="Simple mode",
-			PubmedID="",
+			PubmedID="29040600",
 			Description="",
 			use_by_default=TRUE,
 			heterogeneity_test=FALSE
@@ -182,7 +182,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_weighted_mode",
 			name="Weighted mode",
-			PubmedID="",
+			PubmedID="29040600",
 			Description="",
 			use_by_default=TRUE,
 			heterogeneity_test=FALSE
@@ -190,7 +190,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_weighted_mode_nome",
 			name="Weighted mode (NOME)",
-			PubmedID="",
+			PubmedID="29040600",
 			Description="",
 			use_by_default=FALSE,
 			heterogeneity_test=FALSE
@@ -198,7 +198,7 @@ mr_method_list <- function()
 		list(
 			obj="mr_simple_mode_nome",
 			name="Simple mode (NOME)",
-			PubmedID="",
+			PubmedID="29040600",
 			Description="",
 			use_by_default=FALSE,
 			heterogeneity_test=FALSE
@@ -226,6 +226,14 @@ mr_method_list <- function()
 			Description="Doesn't use any weights",
 			use_by_default=FALSE,
 			heterogeneity_test=TRUE
+		),
+		list(
+		  obj = "mr_grip",
+		  name = "MR GRIP",
+		  PubmedID = "",
+		  Description = "Allele coding invariant regression",
+		  use_by_default = FALSE,
+		  heterogeneity_test = FALSE
 		)
 	)
 	a <- lapply(a, as.data.frame)